Add files via upload
Consider we have rings available in 3 colors denoted by characters "R" , "G" and "B" and we have 9 rods starting from 0 to 9. We were required to find the rod with that had all the 3-colored rings.
   (|)    () - Rings, | - Portion of rod
   (|)    
   (|)
Do you remember that game where you had to toss the ring on the rod so that the rod passes through the center cavity of the ring ? This is it.