Add files via upload
In this problem you were required to add balls to a box in a certain manner and return the max no of balls a box has. The box number for each ball was decided by a number that lied between a lower limit and an upper limit. The box number for each ball was the sum of the digits in the number. For example, ball number 321 would belong to 3 + 2 + 1 = 6 box number.