Add files via upload
The question essentially required us to check whether the words in a text string are typeable or not by checking the characters in a broken letters string .
The broken letters string was a string containing unique characters and the text string was a string with no leading or trailing space and the words in text were seperated by white space. By checking whether the character of broken letter is in word in text or not, we were able to determine the typeability of the word and tell how many words from the text were typeable.