Add files via upload
The question was to return the count of good substrings with length 3. 
Here, a good substring is a substring such that all the characters are unique in the substring and the length of the substring is 3.
We had to output the count of such substrings in the main string with index i to i+3.