Add files via upload
The problem asked us to check whether the strings in the words list were typeable using only one portion of the american keyboard, here, string1, string2 and string3. The approach that I chose was to check whether the set of the strings in the 'words' list were a subset of the set of strings1, strings2 or strings3.