Add files via upload
Alice had n candies and her doctor advised her to eat n/2 candies so as to not get FAT. The array candyType contained the types of candies that Alice had and we had to determine the maximum different type of candies that Alice can eat. Just calculate n/2 and if length of set of candyType is less than n/2 then return length of set of candyType else return n/2.