Create Missing&RepeatingNumbers
You are given an array with length n and elements in the array range from 1 to n. Each element apears exactly once except for one element 'p' which appears twice. One of the elements is also missing from the array that lies in the given range 1 to n that is 'q'. You must find p and q .

https://www.codingninjas.com/studio/problems/missing-and-repeating-numbers_6828164?leftPanelTab=1